requirements.txt
ndspy
pillow